refactor(redesign-plugin-menu): Create Keyboard handler hook specific for PluginMenuModal
